perf: limit open files in generateFSTree (#2458)
